pyproject.toml
[tool.poetry]
name = "tweeran"
version = "0.1.0"
description = ""
authors = ["Andrea Marino <am.marinoandrea@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tweeran"}]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.12.1"
spacy = "^3.4.3"
python-dotenv = "^0.21.0"
praw = "^7.6.1"
pandas = "^1.5.2"
google-api-python-client = "^2.69.0"
google-auth-oauthlib = "^0.7.1"
google-auth-httplib2 = "^0.1.0"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"
vadersentiment = "^3.3.2"
pyyaml = "^6.0"
sklearn = "^0.0.post1"
scikit-learn = "^1.2.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
autopep8 = "^2.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
